app/requirements.txt
python-dotenv
openai
asyncio
aiogram
httpx_socks

